Update settings helpers to check for file existance before overwriting
New helper load_secret, single function which encompasses the other helpers.

+++ ganeti_webmgr/ganeti_web/settings/base.py
@@ -66,21 +66,30 @@ SECRET_KEY_LOC = join(SECRET_DIR, 'SECRET_KEY.txt')
 no_secret_msg = "No secrets in environment variable %s or file %s found.\n"
 
 # Settings helpers
-def load_secret(env=None, file=None, create_file=True, secret_size=32):
+def load_secret(env=None, file=None, create_file=True, secret_size=32,
+                overwrite_file=False):
+    """
+    Helper to retrieve secrets from environmental variables or files.
+    """
     if create_file:
-        secret = get_env_or_file_or_create(env, file, secret_size)
+        secret = get_env_or_file_or_create(env, file, secret_size,
+            overwrite_file)
     else:
         secret = get_env_or_file_secret(env, file)
 
     return secret
 
 def get_env_or_file_secret(env_var, file_loc):
-    """
-    Tries to get the value from the environment variable 'env_var', and
-    falls back to grabbing the contents of the file located at 'file_loc'.
+    """Gets secrets from an environmental variable or file.
 
-    If both are empty, or an IOError exception is raised, this returns None
+    :param env_var: The environmental variable
+    :type env_var: str.
+    :param file_loc: The full path to the file
+    :type file_loc: str.
+    :returns: str -- The secret.
+    :raises: ImproperlyConfigured -- If env_var or file_loc values are None
     """
+
     # Grab the env variable
     secret = os.environ.get(env_var, None)
     if secret is None:
@@ -93,11 +102,22 @@ def get_env_or_file_secret(env_var, file_loc):
             raise ImproperlyConfigured(no_secret_msg % (env_var, file_loc))
     return secret
 
-def get_env_or_file_or_create(env_var, file_loc, secret_size=32):
-    """
-    A wrapper around get_env_or_file_secret that will create the file at
-    file_loc if it does not already exist. The resulting file's contents will
-    be a randomly generated value.
+def get_env_or_file_or_create(env_var, file_loc, secret_size=32,
+                              overwrite_file=False):
+    """Gets secrets from an environmental variable or file. Will create file at
+    file_loc if neither env_var or file_loc return values.
+
+    :param env_var: The environmental variable.
+    :type env_var: str.
+    :param file_loc: The full path to the file.
+    :type file_loc: str.
+    :param secret_size: Size of secret key to be generated for file if
+    being created.
+    :type secret_size: int
+    :param overwrite_file: Overwrite existing file at file_loc if already exists.
+    :type overwrite_file: bool
+    :returns: str -- The secret.
+    :raises: ImproperlyConfigured -- If env_var or file_loc values are None.
     """
     # First check if the env_var or file_loc are set/exist
     try:
@@ -108,9 +128,13 @@ def get_env_or_file_or_create(env_var, file_loc, secret_size=32):
         generate = generate_secret
         secret = generate(secret_size)
     try:
-        # Write our secret key to the file.
-        with open(file_loc, "w") as f:
-            f.write(secret)
+        secret_file_exists = os.access(file_loc, os.F_OK)
+        # If it doesnt exist, then we may write
+        # Otherwise it does exist and we need to make sure we're
+        # allowed to overwrite it
+        if not secret_file_exists or overwrite_file:
+            with open(file_loc, "w") as f:
+                f.write(secret)
     except IOError:
         cannot_create_msg = ("Unable to create secret file.\n"
                              "Try creating the file at %s or setting "
@@ -285,15 +309,6 @@ INSTALLED_APPS = (
 
 ROOT_URLCONF = 'ganeti_web.urls'
 AUTH_PROFILE_MODULE = 'authentication.Profile'
-
-# SECRET_KEY = os.environ.get(
-#     'GWM_SECRET_KEY',
-#     open(SECRET_KEY_LOC).read()
-# )
-# WEB_MGR_API_KEY = os.environ.get(
-#     'GWM_API_KEY',
-#     open(GWM_API_KEY_LOC).read()
-# )
 
 SECRET_KEY = load_secret(env='GWM_SECRET_KEY', file=SECRET_KEY_LOC)
 WEB_MGR_API_KEY = load_secret(env='GWM_API_KEY', file=GWM_API_KEY_LOC)
